requirements.txt
asgiref==3.3.1
autopep8==1.5.5
bangla==0.0.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.6
django-baton==2.1.5
django-extensions==3.1.0
django-resized==0.3.11
django-tinymce==3.2.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.1
idna==2.10
packaging==20.9
Pillow==8.1.0
protobuf==3.15.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
